Added --migrate-db command to migrate the db to new schema
A new column `fichub_id` has been added to the database
schema. This new command creates a new table with the
updated schema, then inserts the data back from the old table since
in sqlite3, the new column is always appended to the end of the
list of existing columns and doesn't allow ADD column_name
AFTER column_name syntax.